refactor: CartActivityLifecycleObserver를 라이브러리 모듈로 이동
