+++ alsonglibrary2/src/main/java/com/example/alsonglibrary2/di/CartActivityLifecycleObserver.kt
@@ -1,14 +1,11 @@
-package woowacourse.shopping
+package com.example.alsonglibrary2.di
 
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
-import com.example.alsonglibrary2.di.AutoDIManager
 import com.example.alsonglibrary2.di.anotations.ActivityScope
 import com.example.alsonglibrary2.di.anotations.FieldInject
-import woowacourse.shopping.ui.cart.DateFormatter
 import kotlin.reflect.full.declaredMemberProperties
 import kotlin.reflect.full.hasAnnotation
-import kotlin.reflect.full.primaryConstructor
 import kotlin.reflect.jvm.isAccessible
 import kotlin.reflect.jvm.javaField
 import kotlin.reflect.jvm.jvmErasure
